task 3/3: Extract QuoteTimer and fee/discount breakdown into components/, trimming SwapScreen.kt to ≤350 lines (#4836)
